+++ hoomd-4.2.1-procf/hoomd/md/EvaluatorPairALJ.h
@@ -1,6 +1,8 @@
 // Copyright (c) 2009-2023 The Regents of the University of Michigan.
 // Part of HOOMD-blue, released under the BSD 3-Clause License.
 
+// ########## Modified by Rheoinformatic //~ [RHEOINF] ##########
+
 #ifndef __EVALUATOR_PAIR_ALJ_H__
 #define __EVALUATOR_PAIR_ALJ_H__
 
@@ -446,7 +448,7 @@ template<unsigned int ndim> class EvaluatorPairALJ
                                 Scalar4& _qj,
                                 Scalar _rcutsq,
                                 const param_type& _params)
-        : dr(_dr), rcutsq(_rcutsq), qi(_qi), qj(_qj), _params(_params)
+        : dr(_dr), rcutsq(_rcutsq), qi(_qi), qj(_qj), _params(_params) 
         {
         }
 
@@ -478,6 +480,22 @@ template<unsigned int ndim> class EvaluatorPairALJ
         {
         return false;
         }
+
+    //!~ Whether the pair potential uses typeid. [RHEOINF]
+    HOSTDEVICE static bool needsTypes()
+        {
+        return false;
+        }
+    HOSTDEVICE void setTypes(unsigned int typei, unsigned int typej) { }
+    //~
+
+    //!~ Whether the pair potential uses timestep. [RHEOINF]
+    HOSTDEVICE static bool needsTimestep()
+        {
+        return false;
+        }
+    HOSTDEVICE void setTimestep(uint64_t timestep)  { }
+    //~
 
     /// Whether the potential implements the energy_shift parameter
     HOSTDEVICE static bool constexpr implementsEnergyShift()